+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## 0.1.4 (2021-03-01)
 
-HTR to latex (add mathpix API)  
+HTR to latex (by mathpix API)  
+Drag and paste image to latex (by mathpix API)  
 
 ## 0.1.3 (2020-12-08)
 
@@ -10,7 +11,7 @@ icon
 
 ## 0.1.2 (2020-12-06)
 
-HTR to latex  
+HTR to latex (by myscript API)  
 Customize buttons  
 CSP script nonce