+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## doing (2022-06-05)
 
-shortcut bind of webview (reference: fork of [joseaeltala](https://github.com/joseaeltala/vscode-markdown-draw/commit/b24a96ac7710a7177461bf7056152c79a81bd97d) and [JL-MORENOCID](https://github.com/zhaouv/vscode-markdown-draw/compare/master...JL-MORENOCID:master))
+Shortcut bind of webview (reference: fork of [joseaeltala](https://github.com/joseaeltala/vscode-markdown-draw/commit/b24a96ac7710a7177461bf7056152c79a81bd97d) and [JL-MORENOCID](https://github.com/zhaouv/vscode-markdown-draw/compare/master...JL-MORENOCID:master))
+
+Collapse svg input element and add toggle button [PR#4](https://github.com/zhaouv/vscode-markdown-draw/pull/4). Thanks [hall](https://github.com/hall).